refactor(cache-admin): migrate CacheAdmin module to inline controller logic
Phase 3 of architecture simplification - CacheAdmin Module Migration:

- Migrated 6 CacheAdmin use cases to inline logic in CacheAdminController:
  - GetCacheStatsUseCase -> getCacheStatistics()
  - ClearCacheUseCase -> clearCache()
  - GetCacheEntriesUseCase -> getCacheEntries()
  - ManualCleanupUseCase -> manualCleanup()
  - GetCacheHealthUseCase -> getCacheHealth()
  - GetRedisHealthUseCase -> getRedisHealth()

- Deleted 6 use case files (~622 lines):
  - GetCacheStatsUseCase.swift
  - ClearCacheUseCase.swift
  - GetCacheEntriesUseCase.swift
  - ManualCleanupUseCase.swift
  - GetCacheHealthUseCase.swift
  - GetRedisHealthUseCase.swift

- Deleted 3 use case test files (~1,017 lines):
  - GetCacheStatsUseCaseTests.swift
  - ClearCacheUseCaseTests.swift
  - GetRedisHealthUseCaseTests.swift

- Updated supporting files:
  - Removed CacheAdminUseCases from UseCaseAccessors.swift
  - Removed cache use case registrations from CQRSServiceProvider.swift
  - Updated ServiceRegistryIntegration.swift validation

Total: ~1,521 net lines removed
