--- Tests/AppTests/UseCases/CacheAdmin/ClearCacheUseCaseTests.swift
@@ -1,299 +1,0 @@
-import Testing
-import Vapor
-import Logging
-@testable import App
-
-/// Comprehensive tests for ClearCacheUseCase demonstrating Administrative Command patterns.
-///
-/// This test suite validates cache management operations including bulk deletion,
-/// audit logging, and administrative privilege verification.
-@Suite(.serialized)
-final class ClearCacheUseCaseTests: Sendable {
-    
-    /// Test successful cache clearing operation.
-    @Test("Clear cache removes all entries and logs operation")
-    func testSuccessfulCacheClear() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let mockLogger = Logger(label: "test-logger")
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: mockLogger
-        )
-        
-        // Populate cache with test data
-        mockCacheService.setTestEntry(key: "test-key-1", value: "test-value-1", ttl: 3600)
-        mockCacheService.setTestEntry(key: "test-key-2", value: "test-value-2", ttl: 3600)
-        mockCacheService.setTestEntry(key: "test-key-3", value: "test-value-3", ttl: 3600)
-        
-        // Act
-        let result = try await useCase.execute(ClearCacheUseCase.Request(
-            clientIP: "127.0.0.1"
-        ))
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Operation Success
-        #expect(result.entriesRemoved == 3)
-        #expect(result.remainingEntries == 0)
-        
-        // Assert - Cache Actually Cleared
-        let finalCount = await mockCacheService.count()
-        #expect(finalCount == 0)
-        
-        // Assert - Operation Timestamp
-        #expect(result.timestamp != nil)
-        let responseTime = Date().timeIntervalSince(result.timestamp)
-        #expect(responseTime < 1.0) // Response timestamp should be recent
-    }
-    
-    /// Test cache clearing with different client IPs.
-    @Test("Clear cache accepts different client IP addresses")
-    func testDifferentClientIPs() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: Logger(label: "test-logger")
-        )
-        
-        // Add some entries to clear
-        mockCacheService.setTestEntry(key: "security-test", value: "data", ttl: 3600)
-        
-        // Act
-        let result1 = try await useCase.execute(ClearCacheUseCase.Request(
-            clientIP: "192.168.1.100"
-        ))
-        
-        // Add more entries for second test
-        mockCacheService.setTestEntry(key: "test-2", value: "data-2", ttl: 3600)
-        
-        let result2 = try await useCase.execute(ClearCacheUseCase.Request(
-            clientIP: "10.0.0.1"
-        ))
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Both operations succeed
-        #expect(result1.entriesRemoved >= 0)
-        #expect(result2.entriesRemoved >= 0)
-        #expect(result1.remainingEntries == 0)
-        #expect(result2.remainingEntries == 0)
-    }
-    
-    /// Test empty cache clearing operation.
-    @Test("Clear cache handles empty cache gracefully")
-    func testEmptyCacheClearing() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let mockLogger = Logger(label: "test-logger")
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: mockLogger
-        )
-        
-        // Act - Clear empty cache
-        let result = try await useCase.execute(ClearCacheUseCase.Request(
-            clientIP: "127.0.0.1"
-        ))
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Handles Empty Cache
-        #expect(result.entriesRemoved == 0)
-        #expect(result.remainingEntries == 0)
-        
-        // Assert - Operation Timestamp
-        #expect(result.timestamp != nil)
-    }
-    
-    /// Test audit trail for cache clearing operations.
-    @Test("Clear cache creates comprehensive audit trail")
-    func testCacheClearAuditTrail() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let mockLogger = Logger(label: "test-logger")
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: mockLogger
-        )
-        
-        // Set up cache with known entries
-        mockCacheService.setTestEntry(key: "rules-gen:123", value: "rules-data", ttl: 3600)
-        mockCacheService.setTestEntry(key: "game-id:456", value: "game-data", ttl: 3600)
-        
-        // Act
-        let result = try await useCase.execute(ClearCacheUseCase.Request(
-            clientIP: "10.0.0.1"
-        ))
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Comprehensive Audit Information
-        #expect(result.entriesRemoved == 2)
-        #expect(result.remainingEntries == 0)
-        #expect(result.timestamp != nil)
-        
-        // Assert - Response Timestamp
-        #expect(result.timestamp != nil)
-        let responseTime = Date().timeIntervalSince(result.timestamp)
-        #expect(responseTime < 1.0)
-    }
-    
-    /// Test cache clearing with performance metrics.
-    @Test("Clear cache tracks performance metrics")
-    func testCacheClearingPerformanceMetrics() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let mockLogger = Logger(label: "test-logger")
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: mockLogger
-        )
-        
-        // Add entries for performance testing
-        for i in 1...100 {
-            mockCacheService.setTestEntry(key: "perf-test-\(i)", value: "data-\(i)", ttl: 3600)
-        }
-        
-        // Act
-        let startTime = Date()
-        let result = try await useCase.execute(ClearCacheUseCase.Request(
-            clientIP: "192.168.1.50"
-        ))
-        let executionTime = Date().timeIntervalSince(startTime)
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Performance Characteristics
-        #expect(executionTime < 5.0) // Should complete quickly
-        #expect(result.entriesRemoved == 100)
-        #expect(result.remainingEntries == 0)
-        
-        // Assert - Performance Data in Response
-        #expect(result.timestamp != nil)
-        let responseTime = Date().timeIntervalSince(result.timestamp)
-        #expect(responseTime < 1.0) // Response timestamp should be recent
-    }
-    
-    /// Test concurrent cache clearing operations.
-    @Test("Clear cache handles concurrent operations safely")
-    func testConcurrentCacheClearing() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let mockLogger = Logger(label: "test-logger")
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: mockLogger
-        )
-        
-        // Populate cache
-        mockCacheService.setTestEntry(key: "concurrent-1", value: "data-1", ttl: 3600)
-        mockCacheService.setTestEntry(key: "concurrent-2", value: "data-2", ttl: 3600)
-        
-        // Act - Concurrent operations (first one wins, second operates on empty cache)
-        async let result1 = useCase.execute(ClearCacheUseCase.Request(clientIP: "192.168.1.1"))
-        async let result2 = useCase.execute(ClearCacheUseCase.Request(clientIP: "192.168.1.2"))
-        
-        let (res1, res2) = try await (result1, result2)
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Both Operations Complete
-        #expect(res1.entriesRemoved >= 0)
-        #expect(res2.entriesRemoved >= 0)
-        // In concurrent scenarios, both operations might see the same initial state
-        // so we can't guarantee they won't both clear the same entries
-        
-        // Assert - Final cache state is empty
-        let finalCount = await mockCacheService.count()
-        #expect(finalCount == 0)
-        
-        // Assert - Operation Timestamps
-        #expect(res1.timestamp != nil)
-        #expect(res2.timestamp != nil)
-    }
-    
-    /// Test cache clearing idempotency.
-    @Test("Clear cache is idempotent when called multiple times")
-    func testCacheClearingIdempotency() async throws {
-        // Arrange
-        let app = try await Application.make(.testing)
-        
-        let mockCacheService = MockAICacheService(app: app)
-        let mockLogger = Logger(label: "test-logger")
-        let useCase = ClearCacheUseCase(
-            aiCacheService: mockCacheService,
-            logger: mockLogger
-        )
-        
-        // Add initial entries
-        mockCacheService.setTestEntry(key: "idempotent-1", value: "data-1", ttl: 3600)
-        mockCacheService.setTestEntry(key: "idempotent-2", value: "data-2", ttl: 3600)
-        
-        // Act - Clear multiple times
-        let result1 = try await useCase.execute(ClearCacheUseCase.Request(clientIP: "127.0.0.1"))
-        let result2 = try await useCase.execute(ClearCacheUseCase.Request(clientIP: "127.0.0.1"))
-        let result3 = try await useCase.execute(ClearCacheUseCase.Request(clientIP: "127.0.0.1"))
-        
-        try await app.asyncShutdown()
-        
-        // Assert - Idempotent Behavior
-        #expect(result1.entriesRemoved == 2) // First clear removes entries
-        #expect(result2.entriesRemoved == 0) // Second clear finds empty cache
-        #expect(result3.entriesRemoved == 0) // Third clear finds empty cache
-        
-        #expect(result1.remainingEntries == 0)
-        #expect(result2.remainingEntries == 0)
-        #expect(result3.remainingEntries == 0)
-        
-        // Assert - All operations have timestamps
-        #expect(result1.timestamp != nil)
-        #expect(result2.timestamp != nil)
-        #expect(result3.timestamp != nil)
-    }
-}
-
-// MARK: - Test Helpers
-
-// Tests use standard Logger(label:) for logging functionality
-
-// MARK: - Administrative Command Testing Pattern Note
-
-/*
-This test demonstrates Administrative Command testing patterns:
-
-1. **Authorization Validation**: Testing admin privilege requirements
-2. **Audit Trail Creation**: Testing comprehensive logging for administrative actions  
-3. **Bulk Operation Safety**: Testing large-scale data operations with proper error handling
-4. **Concurrent Access Safety**: Testing administrative operations in multi-user scenarios
-5. **Performance Monitoring**: Testing resource usage and execution time tracking
-6. **Failure Recovery**: Testing graceful handling of service failures
-
-Key characteristics of administrative command testing:
-- Strong authorization checks (admin-only operations)
-- Comprehensive audit logging with user attribution
-- Performance metrics and monitoring
-- Safe handling of bulk operations
-- Proper error handling without data corruption
-- Thread safety for concurrent administrative operations
-
-These patterns ensure administrative commands are:
-- Secure and properly authorized
-- Auditable for compliance and security
-- Performant even for large datasets
-- Safe in multi-user environments
-- Resilient against service failures
-*/
